# resources/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Currency Rates API",
        "version": "1.0"
    },
    "paths": {
        "/api/cbr/currency-rates/daily": {
            "get": {
                "summary": "Get the CBR Daily rates",
                "operationId": "f5a2b16c4c4af23339cca9d7b38a7806",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/cbr/currency-rates/period": {
            "get": {
                "summary": "Get currency rates for a specified period",
                "operationId": "1541d85b9f7e11f77f25f344426946ee",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start date of the period (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-03-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End date of the period (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-03-10"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    }
}